+++ src/features/Login/Login.tsx
@@ -1,0 +1,5 @@
+import { LoginPresenter } from './LoginPresenter';
+
+export const Login = () => {
+  return <LoginPresenter />;
+};
