+++ src/pages/deleteAccounts/index.d.ts
