+++ src/pages/deleteAccounts/index.tsx
